tests: docker based development/test image and scripts
A docker image that contains all dependencies installed (deb and pip)
replaces the former image. The script part (running Xvfb, ./manage.py
etc.) is delegated to scripts in the bin directory instead. This
minimal image is the basis to run tests, shell interaction, screenshot
generation etc.

The creation of the image is no longer a separate makefile target: it
is always created or updated before running a command in the
container. Checking the image does not need to update does not take
long and the delay imposed when running a command in the container is
minimal.

The bin/dev-shell does something different (for docker build and
docker run) if it runs in CircleCI or not. It would be unecessarily
complicated to unify the two because:

* bind mounting the directory or running under the same user in the
  host and in the container is not required for CI
* the docker container does not run on the same host as the CI
* the CI needs explicit instructions to use the Docker cache

The bind/dev-deps daemons are run from with functions that capture
their output to reduce the output cluter: it is only displayed in case
a daemon fails.

A function to create a demo user with the same credentials as
http://demo.securedrop.club is added to facilitate testing.