Docs: Add note that running tests as root causes trouble
Running the test suite as root will cause a 500 on the source
interface on future visits due to changing permissions of files
that the www-user needs to access. This means that the developer
should be warned that they can not both run the tests and use their
staging VMs for interactive testing.