tests: add test journalist to development environment
The demo user is created with the same user, password and OTP secret
as what is displayed in the online demo. It will allow this function
to be used to update the online demo and people who already have these
credentials memorized in the browser and OTP application will still
be able to use them.

The Journalist.set_password function is not used because it enforces
seven words diceware passwords and the demo password does not comply.
Although it would make for a more educative example to use a seven
words password, it is more difficult to copy/paste, all pages
displaying the existing password would have to be modified and the
users would need to update.