Fixes #3031 type annotation for application config
We have now have `SDConfig` class to provide all the attributes
of `config.py` module. This also removes the need of using any
fake config class, as we can just use a new object of `SDConfig`
for testing.

Becasue of the type check, we have all possible attributes of
`config.py` available by default in `SDConfig`. Right now, the
types are implicit, in future we can make them explicit.