tests: delegate runing lint to a script instead of a Makefile target
The actual command run by the lint docker image is embedded in the
image and the developer does not see it. It is possible to override it
with:

    docker run --rm -ti securedrop-lint:$(git rev-parse HEAD) \
        sudo make typelint

for instance but it is not convenient to remember. The command is
removed from the image and added to the Makefile target. To allow
copy/pasting and modifying the command, the details of the docker
image building and container invocations are wrapped in a script:

    $ make ci-lint
    devops/scripts/dev-shell-ci sudo make --keep-going lint typelint
    ...