Release script: Update script for new release candidate versioning
Previously, we used the version string 0.x.y-rcN for the Nth release
candidate for 0.x.y. Unfortunately, this rc version string was
considered a higher version than the actual release.

In 0.5.2 and beyond, we use the version string 0.x.y~rcN for the Nth
release candidate for 0.x.y. This versioning scheme is correct, in
that Debian and Ubuntu recognize 0.x.y as a higher version than
0.x.y~rcN.

This commit updates our release script to create valid tags for
release candidates. Since the character "~" is not
a valid character in a git tag, without this change the script
will fail.

Issue: https://github.com/freedomofpress/securedrop/issues/2950