build: re-create the builder docker image when Dockerfile changes
It takes longer but works all the time. Uploading an image to a docker
repository to speed-up the build requires an additional manual step
and access to the docker repository.

A possible optimization (would be a few minutes faster) could be
implemented by storing docker layers, similar to what is done for the
Dockerfile used when testing.