group and simplify iptables INPUT rules
The rule:

-A INPUT -p tcp -m state --state ESTABLISHED,RELATED -j ACCEPT

shadows all other TCP rules.

The rule:

-A INPUT -i lo -p all -j ACCEPT

shadows all other TCP/UDP rules for lo, regardless of their state.

They are replaced by one rule, for both udp and tcp:

-A INPUT -m state --state ESTABLISHED -j ACCEPT

There is no need to account for the RELATED state because no protocol
for which it would be useful is allowed (ICMP, ftp etc.).

The remaining INPUT rules are left in place only when they explicitly
permit NEW connections on a given port, others are removed.

A new INPUT rule is added to allow new connections to SSH from the
user running the tor daemon.