Updates config tests for new kernel image version
Simple version bump. This *should* be the only string that need change
in the config tests for a new kernel, but CI shall light the way if not.
Perhaps we should be explicitly checking for the version of the
`securedrop-grsec` metapackage in the config tests. Really what we care
about is the version of the running kernel, so sticking with that for
now.