securedrop-admin: symlink ancient .venv to admin/.venv
Prior to securedrop-admin moving to the admin directory, its
virtualenv directory was in the .venv directory at the toplevel of the
git tree. It is now in the admin/.venv directory instead.

For the regular admin following the instructions from the
documentation the existence of the ancient .venv will not be noticed
and can be ignored.

However the more knowledgeable SecureDrop admin who manually source
the .venv will be confused. We remove the .venv and symlink the new
admin/.venv so the change is transparent.