Preserves 3.14.79 kernel during upgrade
Our use of cron-apt includes autoclean and autoremove flags. Unless we
declare an explicit dependency on the 3.14.79 kernel image package, it
may be cleaned up. That's ultimately desirable, but would prevent orgs
that have trouble with the new kernel from falling back to a known-good
kernel. Let's preserve the old image as a fallback for now.

In subsequent releases, we can omit the dependency on 3.x, and simply
require a single kernel version, as we've done in the past. For now 0.6,
let's provide rollback capability, since we cannot test on all hardware.

(cherry picked from commit 97674f3d31638beabf2df99499861939b89c1702)