tests: delete none needs modal instead of obsolete #delete-selected
* click "delete selected"
* wait for the modal
* click the modal so the error is displayed

The test would have failed because the #delete-selected no longer
exists. But it still exists, only in the context of a modal which is
hidden. Because of that it was able to find the element and click it,
with no side effect which was the intended purpose.

However the screenshot was wrong and missing the warning explaining
nothing was deleted.

(cherry picked from commit ed338e852bfc9b5724bfa188f4760496792829b5)