Tails: Run updater GUI on boot and only when SD updates are needed
This commit runs check_for_updates on boot, and then runs the
updater GUI if updates are needed. If updates are not needed,
then nothing will happen.

Note: in order to run this, the venv must already be set up, as the
journalist/admin will run `./securedrop-admin tailsconfig` in
order to update this on-boot hook. So assuming that the venv
is configured in the on-boot hook is a safe assumption.