Journalist GUI: If a failure occurs, return user to main GUI window
This is such that they can try again. For example, if they just fail
to type in the sudo password, they might want to try again.
If they get a scarier message they _can_ try again, though the
strings warn the user in such a case just to go to their
administrator.