tests: capture l10n source strings when taking screenshots
If the PAGE_LAYOUT_CAPTURE_L10N environment variable is set,
capture the source strings from the latest request (either from the
source or the journalist interface) and save them in a JSON list with
the same basename as the screenshot.

This list of source strings can later be used to figure out which
screenshot contains which source string.

We are using an environment variable PAGE_LAYOUT_CAPTURE_L10N instead
of a pytest argument because it needs to be used in the setup()
function and there is no convenient way to access the pytest arguments
from this function.