Fix playbook when user toggles ssh_tor -> localnet
This is a rare scenario that I'm addressing and this isnt the smoothest
logic in the world, but it does the trick. All of our roles are
currently designed for the opposite scenario (transistioning from ssh
localnet -> ssh over tor), going the opposite way is tricky. So I added
a minimal amount of conditional logic and have it bail early with a
friendly user message. I tried to do a message and a meta `end_play`
event instead so it wouldn't show as an "error" but I was unsuccessful
to get that working as documented