Explicitly disable NFS in upgrade scenario
This is super annoying that I have to keep messing with this option. I
suspect its a weird KVM + vagrant issue :shrug: Don't have time to dig
deeper and this fixes it along with my fork of upstream molecule vagrant
module. Both changes of which, I have open PRs to try to get merged in
so we dont have to maintain this divergence.