Add CI job for updater GUI tests
Without installing libqt5designer5, Circle CI is unable to load the
platform plugin xcb [0]. It is installed, but it will not successfully
load until this dependency is added. This will cause a failure.

The reason I elected to install the dependency in a run step and
not just make a custom Dockerfile is that making a custom Dockerfile
seemed like overkill for installing a single apt dependency.

[0] https://forum.qt.io/topic/30257/solved-failed-to-load-platform-plugin-xcb/2