Add a Code of Conduct to the SecureDrop project
Resolves #3251.

A code of conduct is essential for any substantial free software
project, and SecureDrop is no exception. This code of conduct will set a
standard for how contributors should conduct themselves in relation to
SecureDrop.

This was originally spearheaded by myself, and made possible by the help
of several collaborators, including @eloquence, @redshiftzero, and
@olivemartini.

This code of conduct complies with the GitHub recommended file format.