l10n: revert to using an old translation
https://github.com/freedomofpress/securedrop/pull/3379

reverted a change at the last minute to resolve an upgrade problem
prior to the release of 0.7.0. This contains the following user facing
change:

 #: journalist_app/forms.py:62
-msgid "You can only upload JPG/JPEG or PNG image files."
+msgid "Upload images only."
 msgstr ""

Which was translated in the past and the translation was preserved,
allowing us to restore it without asking the localizers.

Relates to: #3379

(cherry picked from commit eab19aa7d6cbae681ee4f35a845ed9aef949a676)