Fixes for vagrant_packager + updated molecule
With the bump to molecule, the ephemeral directory spot has changed. For
now, lets utilize the old directory space since its already git/docker ignored. We
just need to ensure that we intentionally create that directory.

There was also a change in how the vagrant boxes are named. The
`.molecule` prefix is gone and replaced with the scenario name.