Security improvements to logic in securedrop-admin update
* We should check out the branch only once the signature verifies.
This way if the signature does not verify for some reason, the user
will have the previous branch checked out. In the release notes
going forward, if we ask users to manually check out another branch,
we should follow this procedure (previous to this, we had users verify
the signature after checking out the tag).

* We should check the entire fingerprint for each update and return
a signature verification error if not.