Install cryptography==2.0.3
Latest pyca/cryptography requires setuptools>18.5, which we do not
have on prod installs. In order to make this transition in the next
minor release however, we can temporarily pin to a slightly older
version of pyca/cryptography, such that we can install it using
the version of setuptools that we have.

For the Docker dev env, we need libffi-dev installed to pip install
cryptography.