ansible: rules_v4 does not need to inspect unused ipv4 details
The iptables rules are only used when enable_ssh_over_tor is not
set. Save ansible the trouble of examining the details of the ipv4
address and interface when the result is not going to be used.

The main benefit is also to workaround a border case which is not
properly handled by ipaddr:

   ipaddr('184.20.18.54/255.255.255.255')

although valid fails with

   AnsibleFilterError: ipaddr: unknown filter type

This border case may be found in colocations where IP addresses are
distributed individually as /32 instead of being part of an allocated
subnet. It is however unlikely to be found in test environments or in
setups where the machines are allocated IPs in a private subnet.