Fix skip on latest 2.4 ansible, mv build-deb-pkgs
Two things done here:
* When we bumped ansible/molecule versions the previously working logic
targeting a host per role was barfing on the delegation. Does this make
sense .. maybe they just "fixed the glitch" that allowed me to do that
previously. :shrug: dont have time to investigate but implemented a fix
on the conditional.

* Moved the build-deb-pkgs playbook from ansible folder to molecule.
Why? Its worthless on its own without the standup of the containers so
it might as well live inside the molecule folder and save some space.