QA loader: Sources should be viewable in the journalist interface
To do so, they must have pending set to False (by default it is
True, and such sources will not appear in the journalist interface),
as well as a filesystem_id that is not None. Any source that has
a filesystem_id of None will cause a 500 server error, so this is
a reasonable assumption to make for existing instances.