journalist notifications: allow 30 minutes variance in reboot times
The race condition goes as follows:

  - machine reboots
  - notification sent
  - machine reboots 24h later but reboots 1 minute faster than the previous day
  - notification is sent 23h59 minutes after the last one and suppressed

The notification will be sent the next day but the information it
contains will not reflect what happened in the past 24h because the cron job

   manage.py were-there-submissions-today

updates it daily, one hour before the machine reboots.

To resolve this race condition (or make it extremely unlikely) we
allow for notifications to be sent at most every 23h30.

Case 1: machine is slower to reboot the next day:

  - machine reboots
  - notification sent
  - machine reboots 24h later but reboots 5 minutes slower than the previous day
  - notification is processed 24h05 minutes after the last one, it is
    more tha 23h30 and therefore the notification is sent

Case 2: machine is faster to reboot the next day:

  - machine reboots
  - notification sent
  - machine reboots 24h later but reboots 8 minutes faster than the previous day
  - notification is processed 23h52 minutes after the last one, it is
    more tha 23h30 and therefore the notification is sent