journalist notifications: today's submissions count 1h before reboot
If the sysadmin chose to:

a) activate journalist notifications
b) modify the reboot time

It is possible that the submissions count is older than expected. For
instance, if the submissions count is done at midnight and the machine
reboots at 11pm, the journalist will receive a submission count that
is about 24h old and does not include submissions received in the past
23h00.

The information sent to the journalist is correct, only it is older
than one would expect.

The submissions count is updated one hour before the machine reboots
time so it is no more than one hour old, regardless of the reboot time
chosen by the admin.