Add postinst to securedrop-grsec to preverve boot order
For SecureDrop instances running 3.14.79, the postinst hook will, for instances where kernels are rolled back, specify the canonical name rather than ordinal name for the default kernel to boot. In cases where instances are rolled back, an update to 4.4.135 will likely result in the instance not booting due to the position of the kernels changing in the grub configuration.