Update PR template to add checkbox for test plan
In the retro for 0.7.0, we discussed being more rigorous about
test plans. Part of this is team members writing test plans, ensuring
that reviewers carefully evaluate the test plan and whether it covers
all important cases, and multiple people running through the test
plan. Part of this is ideally the PR submitter writing a test plan
and validating that their PR passes the test plan prior (just to
reduce back and forth a bit).