Molecule builder tests: Skip test_ensure_no_updates_avail in CI
This test checks for security updates in the build container.
While this is good, security updates are needed several times a
week, and sometimes multiple time in the same day. It is critical
to update when we are building test or production debian packages,
and otherwise we should not block CI.