ossec: fix journalist notification border case
When .procmailrc changes, the pattern triggering this script will match
itself. Just return and ignore the match.

In install_files/ansible-base/roles/postfix/files/procmailrc the following:

    :0 B
    * /var/lib/securedrop/submissions_today.txt
    |/var/ossec/process_submissions_today.sh

was introduced in 0.7.0. It is only installed when running
securedrop-admin install. Since this file is monitored by OSSEC, it
will trigger an alert with the following:

    Integrity checksum changed for: '/var/ossec/.procmailrc'
    Size changed from '162' to '261'
    What changed:
    6,8c6,13
    < :0 c
    < *^To:.*root.*
    < |/var/ossec/send_encrypted_alarm.sh
    ---
    >
    > :0 B
    > * /var/lib/securedrop/submissions_today.txt
    > |/var/ossec/process_submissions_today.sh
    >

Which matches the rule and therefore triggers the script which fails
to find the expected submissions count.

(cherry picked from commit ba954805bc38033c440210471721d1d7c00f85fc)