Journalist API: Refactor error handling
This commit makes several changes to error handling on
the journalist interface.

First, CSRF protection is not necessary for a REST API that is not
using cookies for authentication, which this one is not (we
pass a token in a HTTP header). We exempt only the api blueprint
views from CSRF protection, and leave it as is for the remainder
of the journalist application.

Second, we use blueprint-level handlers so that we can more easily
present the relevant response to the user: an informative JSON
error response (i.e. one that contains details of why the request
was not valid) or an HTML page that is rendered in the style of
the rest of the journalist application.

Note that blueprint level 404 and 405 error handlers are not respected,
so we add logic to use the API's error handlers to return nice
error messages to the user.