Renames Qubes default username -> "sdadmin"
Clarifying that this is a user account for a human administrator, rather
than a machine account used for running the SecureDrop application.
We've used "sdadmin" in other contexts, as well, such as the default
vars for configuring an instance via a Tails Workstation.

It doesn't really matter what this value is, but it will be hardcoded in
other aspects of the Qubes staging environment, and so any changes down
the road will require substantial developer intervention in order to
accommodate. So "sdadmin" is the most conservative approach, and
hopefully the most future-proof, as well.