Documents dev env setup in Qubes
Straightforward, mostly the same as the existing Debian 9 install docs,
with a few dom0 commands, as is to be expected in Qubes. We recommend a
StandaloneVM in Qubes, because otherwise the developer must:

  * manage a custom template for the dev AppVM, or
  * reinstall docker on every boot (no, thanks)

The compromise here is additional overhead for the StandaloneVM, but
that's well worth it for a smooth dev experience.