Reuse ansible.cfg in Qubes Molecule scenario
We want the pipelining option in particular, but in general it's sound
practice to reuse the config so we're testing what's running in prod in
as many different contexts as possible. The dynamic inventory statement
in the ansible.cfg is not used, due to precedence of the Molecule
inventory file.