source: do not set last_updated on key generation
The last_updated field is updated every time a submission is
sent. Also updating it when the source key is generated has a
different meaning and creates a problem.

A: there is enough entropy and the key is created when a message or
   document is submitted. The last_updated field is set twice, with
   the same value, give or take a few seconds.

B: there is not enough entropy and the key is created long after the
   initial document or message was submitted, when the source
   reconnects using the same passphrase. The last_updated is set and
   that confuses the manage.py::were_there_submissions_today which
   assumes the field is only set when a submission is received.

Since no code logic relies on the last_updated field being set when
the key is generated, the async_genkey is modified to no longer set
the last_updated field.