Creates new Molecule scenario: libvirt-staging
Intended to replace the Vagrantfile-based staging environment.
As a downside of the move from Vagrant to Molecule, we'll lose the
ability to detect automatically a suitable provider for the Vagrant
driver. That's acceptable: users of libvirt already set
VAGRANT_DEFAULT_PROVIDER=libvirt to override the virtualbox default used
by Vagrant.

The major benefit of the Molecule transition is that it opens up testing
against multiple *platforms*, e.g. Trusty and Xenial, in the near
future. Ideally we can also cut the wrapper logic in
`testinfra/test.py`, since Molecule integrated with Testinfra just fine
by itself.