Converts reboot task to role
In the AWS staging Molecule scenario, we've already implemented custom
reboot logic. For the libvirt-staging scenario, we'd have to do so as
well, given that the relative-path import logic for the reboot task list
only honors the symlink *source* for playbook, rather than the target.

Ansible handles roles much better, with ANSIBLE_ROLE_PATH, so converting
the task logic to a full role resolves import errors when running the
staging playbook in different Molecule scenarios.