source: do not set last_updated default to now
The last_updated field is used to reflect:

* When a new submission is sent
* When the key of a flagged source is created

It should not be set when the source record is created otherwise it
will be reported a an activity worthy of the journalist attention. But
when a new source is created (and before any submission is sent), the
source does not even show (it is pending) and the journalist cannot
notice any difference.