tests: db_helper submit must update last_updated
When the db_helper::submit function creates new submissions for a
source, it must also set the last_updated field in the same way
Storage.submit does.

This difference between db_helper::submit and the function it is
supposed to mimic was of no consequence so far because it was called
shortly after the source is created and last_updated default was the
current time.

Now that the last_updated default is None, some tests that rely on
last_updated when downloading submissions broke because it did not
contain a valid date although submissions are present.