Temporarily ignore safety check for pyca/cryptography
This ignores the safety check failure for CVE-2018-10903. As the flaw is in AES-GCM (SecureDrop app server does not make use of AES-GCM), and that updating the dependency may require updating to Xenial or potentially introducing further risk, let's set safety to (temporarily) ignore this vulnerability. See issue #3677 for more information.