Uses Molecule verify step for AWS scenario
Begins the process of overriding the env vars for staging.
Includes a small change to the tor-apt test, for Python2 compatibility
(otherwise the test can't be run locally).

More work required on balancing the side-effect logic. We'll still need
side-effect to run the app tests, we can run the verify step after
side-effect.

Includes some flake8 linting, since via Molecule, we're now linting the
entire testinfra directory.

Updates JUnit test file handling

The paths to the helper scripts changed a bit, so updating them here. We
had some skipped logic in the side-effect which has no been removed, so
we don't have to maintain it any longer.

Moved the junit logic out of the side-effect since the testinfra tests
are run after the side-effect (and therefore the XML file wouldn't have
existed yet).