Journalist API: Provide journalist UUID in API responses
Now that the conversation history is exposed to journalists, we
need some way for clients to keep track of _which_ journalists
have sent a particular reply to a given source. While this
information is tracked via the journalist_id associated with a
given reply in the database on the server, this also needs to be
tracked on the client side. We cannot do this via the username
as this can be changed. Instead, we need a UUID per journalist which
is mapped to the username, and expose that UUID using the API.

Using this UUID, we can track which journalist has responded to a
given source. By providing both the journalist username and UUID
in API responses, we can also update the username of each journalist
server-side if it is changed on the server.