Creates Molecule scenario for evaluating Xenial
Exactly the same as the other staging scenarios, except for a few
variable overrides. Duplicating the logic here for libvirt and
virtualbox; have not implemented Qubes support, since that would require
building a base VM from a Xenial ISO. We can add Qubes support as we get
closer to stabilizing the Xenial logic.