Provides overrides for trusty/xenial apt channels
In both the FPF apt repo config and the tor apt mirror, the lsb.codename
logic will automatically target Xenial when referencing the remote
apt repos. That won't work without backend changes to the apt server,
which we're not ready to make. Let's preserve the current codename-based
behavior, but support overrides, so we can selectively force "trusty"
while evaluating support for Xenial.