QA data loader: Set default multiplier at 25
Using a multiplier of 100 in this script adds over 5000 sources to
the database, each with many submissions and replies. At the scale
that SecureDrop is currently at, e.g. without pagination on the
journalist interface, it is highly unlikely any organization has this
many sources saved. Even if they had a submission saved from every
day for the last several years, they'd still have less than 5000.

The large number of sources can also increase the chance that there
are timeouts on the journalist interface, due to the lack of pagination.

As such, we should make this more realistic, e.g. a multiplier of 25,
which adds 1250 sources to the database.