Bumps CI timeout on app tests 10m -> 20m
The default timeout option for a given CircleCI task is 10 minutes. If
no output is printed to stdout during that period, the build is
cancelled and marked as failed. Since we're running the app tests
via an Ansible wrapper, stdout is buffered, triggering the timeout.

Let's bump that value to 20m to give the tests a chance to report
results. See docs here:

https://circleci.com/docs/2.0/configuration-reference/