CI on release branch: Remove customization of Tor apt sources
This removes now unnecessary customizations only on release branches.
We no longer should remove the tor_apt_freedom_press.list as of #3189
that is the primary Tor apt repo.

We now add test Tor packages to apt-test.freedom.press, so they will
be automatically installed provided they are more recent versions that
what is live on tor-apt.freedom.press (which is the case for pre-release
QA of Tor packages).