Update cwd for show-refs command
When executed in the context of the updater gui, the working directory
for the updater script that was called was `/`, resulting in an error stating that the current working directory (or any of its parents) was a git repository.
This fix will ensure the SecureDrop updater GUI will not erroneously return a signature verification failure.

(cherry picked from commit 04586f25b121fb8fd3f865fcaee30bb552708c95)