Updates tor config tests for Xenial support
Mostly addressing the sysv vs. systemd style service declarations. Under
Xenial with systemd we can use the builtin "service" module to confirm
running and enabled, which is a bit simpler. Duplicated the tests to
return true on the non-target platform, since we cannot skip via pytest
without knowledge of the target platform (which requires connecting, and
therefore executing the test).