Add make helper cmds surrounding upgrade testing
I'm not sure the Makefile is the best place for it... this file is
getting out of hand - but for now... yep.

By default lets start the upgrade scenario from the latest stable..
obviously this will be problematic when we release and there is a lag in
publishing the latest baseline VMs but I dont see this as a huge issue.
Just gotta get better at updating the boxes post release.