Hacky work-around for ephemeral directory move
At some point in molecule dev, the ephemeral directory was moved from
the local scenario dir + `.molecule` to a path in
/tmp/molecule/${git_dir}/${scenario_dir} . This is a welcome change
EXCEPT that there could be an edge case where someone clones in
securedrop without the default `securedrop` dir path. This is
problematic because the molecule.yml will not take dynamic interpolation
AND its the only spot i could figure out to specify the ansible roles
dir.

So I wanted to take advantage of the new location of the ephermeral
directory while also having a hard-coded roles_dir that wont change. The
work around here is moving the git clone to the official molecule
ephemeral dir, but using the legacy spot to make a roles symlink. There
is some DRY improvements that can obviously be made but since this is
already a crazy commit I tried to hold my self back from more
refactoring.