Boot shellcheck linter to its own script
Two part reason for doing this:
* logic in Makefile was getting too large .. its better to manage this
type of stuff in a dedicated script

* I had to make significant changes as part of the `match CI lint to dev
lint`

One thing you might note is that I dropped use of perl for sed+awk. Why
did I do this? I saw some weird behavior with the perl logic locally -
while I was debugging and I realized as soon as I changed it that I
uncovered a bunch of failing tests.

I also wanted to drop the weird call docker from inside docker logic we
had going on.. that was going to continue to get hairy as I refactored.