Update Admin Guide with how to configure locales
Resolves #3846.

Documentation on how to add additional locales was easily available for
new instances. However, existing admins would likely not find these
docs, creating a pain point for support. In the interest of DRYing
support a bit, adding a section on how to add locales explicitly will
make it easier for SecureDrop admins in the short term.