Override maild.full_subject in ossec server config
Setting maild.full_subject=1 for ossec >= 3.0.0 will ensure alert does not appear in the email subject line (and thus unencrypted as it is sent to the admin).

Setting it in local_internal_options will ensure that it is not squashed by changes in internal_options that is shipped as part of the ossec-server package.