Updates docs for GCE (rather than AWS)-based CI
Cleaned up the docs a bit to reference the new script paths, as well as
excise outdated info such as "we don't run kernel tests in CI". We do!

The Makefile targets aren't necessary, but do provide a bit of
convenience to developers.

Updated the "ci-go" logic in particular, in order to preserve the "Don't
run staging CI if branch name starts with `docs-`" logic.