Removes uid-owner restriction from apt iptables rules
Under Trusty, apt calls run as the root user. Under Xenial, however, apt
calls run as the `_apt` user. We must ensure that the firewall rules
permit apt updates on both systems, before and after the upgrade from
Trusty to Xenial. We cannot add references to `_apt` under Trusty, since
the user doesn't exist in the passwd file, causing the iptables-restore
action to fail.

Let's simply remove the uid-owner restriction on the apt-specific
iptables rule. By this point in the iptables logic, non-whitelisted traffic
from both the `debian-tor` and `www-data` users hsa been dropped (via
jump to the LOGNDROP chain). As such, it's fine to loosen the
specificity of the iptables rules somewhat, in order to ensure a smooth
transition.