Adds upgrade testing box for 0.11.0
We missed updating the upgrade boxes during post-release procedure for both
0.10.0 and 0.11.0. Since 0.11.0 is still the most current release, it was
rather straightforward to prepare a box. Catching up with 0.10.0 was trickier,
however, since the box preparation logic assumes that whatever version
is posted to the prod apt repo is the correct one to prepare for; no longer
the case. As such, I'm omitting metadata for the 0.10.0 version.

Since I did prepare and upload those boxes before realizing the 0.10/0.11 version
discrepancy, here are the checksums of what's stored in S3:

  {
    "version": "0.10.0",
    "providers": [
      {
        "name": "libvirt",
        "url": "https://s3.amazonaws.com/securedrop-vagrant/app-staging_0.10.0.box",
        "checksum_type": "sha256",
        "checksum": "2dda15d0f2557dd434dfba4cf0b3ab58b30666ac0277cd8d32b973d954b54264"
      }
    ]
  }

  {
    "version": "0.10.0",
    "providers": [
      {
        "name": "libvirt",
        "url": "https://s3.amazonaws.com/securedrop-vagrant/mon-staging_0.10.0.box",
        "checksum_type": "sha256",
        "checksum": "6e281590a7dfd7753136b7c6505c8dd1ebc4d7f8207c890a80db4871d6ac08b7"
      }
    ]
  }