Implements debian/rules for securedrop-app-code
The top-level `debian/` dir is used for source packages, and via
`debian/rules` will generated `DEBIAN/` dirs for binary packages.

Declaring the deb package compatibility of "5" here, which is the lowest
compatibility version supported by debhelper. We were previously using
inferred compat of "1", given that it was undefined.

Moves changelog for securedrop-app-code

The `dpkg-buildpackage` logic assumes that the package changelog exists
at a specific location. The filepath doesn't appear to be overrideable
via a CLI flag, so moving it. Requires corresponding update in the
update_version logic.