Updates Xenial builder image hash
Use upstream images that contain a fix to the apt vulnerability,
CVE-2019-3462, which was recently patched in develop (before the Xenial
builder image existed).

The `tzdata` dependency is required for building OSSEC packages.

Not storing `python-requests` in the image file, but rather pulling in
via Ansible logic, because including it generated a pyssl error.
We already declare the dependency in declared in the OSSEC Ansible logic
for building, so it's not strictly necessary in the container image,
but it would be nice to speed up builds somewhat.