Updates dev docs with Xenial-specific language
Touches up the package build logic with explicit mention of the Xenial
targets. Also removes several of the known problems with Xenial, given
that we've addressed them as part of recent sprints. Leaving in the
config test verification mention, since we have a separate ticket
tracking that (3964).

Have not made mention of prod envs for Xenial, since we have more work
to do on that front yet.