Removes tor-apt repo from securedrop-config
As part of the 0.12.0 release, we're consolidating the separate tor-apt
repo into part of the main apt.freedom.press repository. As such, we
must clean up references to the old tor-apt repository on disk.
Accordingly, increments the version of the `securedrop-config` package.

Includes updates to relevant Ansible logic, for backwards compatibility,
as well as config tests, which aim to ensure that the old tor-specific
apt repos are not configured anywhere on the system.