Fetches upstream Tor debs via Molecule
Uses another Molecule-based build scenario that fetches the Xenial tor debs
for inclusion in the FPF-maintained apt repository. The debs are fetched into
the same `build/xenial/` dir as the SD debs. There's no corresponding logic
for fetching Trusty debs, since Tor has stopped backporting tor for trusty.

The major change here is the update to the Tor Project public key, which
has expired since we last updated the pubkey stored in this repo. We haven't used
the Tor pubkey directly since switching to an apt mirror for distributing tor debs.

As of 0.12.0, we plan to use a single, consolidated apt repo, to distribute both
SecureDrop and Tor deb packages.