Disables strict xfail on PaX flag config tests
The "strict=true" pragma on pytest xfail means that the config test
suite will exit non-zero if the tests pass. We added strict=True in order
to understand better the problem of PaX flags not matching, and it
appears we've confirmed that the age of installed packages is sufficient
determinant. That is, if apt such as grub are upgraded as part of the
install, then PaX flags set by the grsecurity role will be clobbered
very quickly.