ci: move trusty app tests to nightly job
Beginning with 0.12.0, news organizations will be migrating to
Xenial, and all new installs will be on Xenial. Thus, we should
move the Trusty app test job to a nightly job, and monitor it
for failures until Trusty is deprecated in the following months.

This also reduces total CI time per PR, as the trusty job took 20+ minutes.