Adds regression test for conffiles in app-code
The recent migration to debhelper caused the AppArmor profiles to be
considered conffiles, which breaks our ability to ship unattended
updates to those files. We want the AppArmor profiles to be exactly as
we configure them, including any updates over time, so we cannot permit
the conffiles classification.

This test ensures that the `securedrop-app-code` deb package contains
only the explicitly whitelisted conffile of the default logo location.
Any additional conffiles, automatically classified via debhelper or
otherwise, will cause the test to fail, to guard against regressions.