ci: use YAML anchors to DRY up common docker layer caching steps
This commit only DRYs up the xenial docker layer caching steps
because the trusty job will soon be going away (and because I
only want to use the xenial docker layer cache for the translation
job).

Due to an incomplete implementation of YAML anchors on CircleCI [0, 1],
this commit adds one anchor per step.

[0] https://discuss.circleci.com/t/allow-bundling-several-steps-e-g-by-supporting-nested-steps/15339
[1] https://discuss.circleci.com/t/yaml-anchors-implementation-incomplete/16916/4