Makes upgrade box naming more explicit
Storing the version string and platform directly in the box filename
now. The docs for maintaining the "upgrade" boxes have long included
a manual step to rename files with the version number; that's now
done automatically.

The addition of the platform in the filename is required to support
multiplatforms, specifically Trusty & Xenial. Eventually we'll drop
support for Trusty altogether, but for now we must maintain both.