Remove unused packages
Wireless-related packages are not required for SecureDrop App and Mon servers. `gcc` and dpkg-dev` however are required, as they are listed dependencies for `libgcc1`, a dependency for very large number of packages including `apt`. These packages are installed by default in Trusty, and will remain if an instance is upgraded to Xenial.

(cherry picked from commit 1517e28b4549941153fa521bc7e33470b48cde31)