Adds prepare-servers role to upgrade VM playbook
The prepare-servers role was added to the main prod/staging playbooks,
but wasn't added to the derivative playbook used for building the
"upgrade" Vagrant boxes. Ideally we'd have all role includes
conditionally set, so that we can reuse the same playbook in all
contexts and disable specific roles via vars per scenario.