Resolves #4100
you don't have an interactive prompt in the gui-updater, so you
shouldn't use a tty and --batch or --no-tty is meant for such tings.

--no-tty
     Make sure that the TTY (terminal) is never used for any output.
This option is needed in some cases because
     GnuPG sometimes prints warnings to the TTY if if --batch is used.

--batch
Use batch mode.  Never ask, do not allow interactive commands.

The test, which uses a subshell, is looking for errors on the cli, this
should be changed.