Updates vagrant-package wrapper script
Now uses two blocks, first for Xenial, second for Trusty,
to iterate through the base box creation logic.

Renamed the file slightly to confirm to the other scripts
in `devops/scripts/`.

Updates logic in `package.py` script to detect whether it's running
against Trusty or Xenial, based on the calling scenario. (The script is
symlinked into the Trusty scenario, so we can detect the invocation
path.)