Waits longer for reboots in Xenial packaging
Encountered reliable timeouts on privilege escalation prompts after the
VMs had been instructed to reboot, on the gather_facts step prior to
shutting the VMs down for packaging. Tried the following, in order:

  1. Reused the reboot-and-wait logic from the site config.
  2. Bumping the gather_timeout value (as we do in the debian package build
     logic scenarios).
  3. Added mandatory pause of 30s post-reboot.
  4. Added mandatory pause of 60s post-reboot.

When the final change was added, the problem was reliably resolved. It's
possible we can drop the gather_timeout bump, but given the wall time
involved in testing, I'm deferring further improvements to a later date.

If you're reading this message, consider that ControlPersist files may
be relevant to the timeouts. I haven't confirmed that assumption, but
it's possible that the long pause post-reboot has the side-effect of
allowing the active ControlPersist files on the controller to expire,
forcing a new connection.
FIX: cleans up reboot in packager scen